Update pyspi test profile analysis notebook
Previously was testing/profiling pyspi SPIs on sliced example time
series of target samples and features. But this is inconsistent with how
we will compute pyspi for the actual analyses. Instead, now test/profile
in the same way we will compute later, i.e. using full time series with
a target parcellation and temporal pooling.

The outcome however is mostly the same.